requirements.txt
allure-pytest==2.13.5
pytest==8.1.1
selenium==4.22.0
